+++ src/Birski/utils/Configs.java
@@ -4,11 +4,11 @@ import java.awt.*;
 
 public class Configs {
 
-    public static final short DRAW_BOARD_SIZE = 500;
+    public static final int DRAW_BOARD_SIZE = 500;
 
-    public static final short RECTANGLE_SIZE = 5;
+    public static final int RECTANGLE_SIZE = 5;
 
-    public static final short NUMBERS_OF_RECTANGLES = 100;
+    public static final int NUMBERS_OF_RECTANGLES = 100;
 
     public static final int COLOR_MAX_VALUE = 255;
